MFB: Fixed bug #44333 (SEGFAULT when using mysql_pconnect() with client_flags)
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@254239 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/mysql/php_mysql.c
@@ -600,7 +600,7 @@ static void php_mysql_do_connect(INTERNAL_FUNCTION_PARAMETERS, int persistent)
 	int  user_len, passwd_len, host_len;
 	char *hashed_details=NULL;
 	int hashed_details_length, port = MYSQL_PORT;
-	int client_flags = 0;
+	long client_flags = 0;
 	php_mysql_conn *mysql=NULL;
 #if MYSQL_VERSION_ID <= 32230
 	void (*handler) (int);
